+++ src/main/java/org/opensearch/knn/index/query/KNNQueryBuilder.java
@@ -65,8 +65,8 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
     private final String fieldName;
     private final float[] vector;
     private int k = 0;
-    private Float max_distance = null;
-    private Float min_score = null;
+    private Float maxDistance = null;
+    private Float minScore = null;
     private QueryBuilder filter;
     private boolean ignoreUnmapped = false;
 
@@ -99,7 +99,7 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         if (k == null) {
             throw new IllegalArgumentException("[" + NAME + "] requires k to be set");
         }
-        validateSingleQueryType(k, max_distance, min_score);
+        validateSingleQueryType(k, maxDistance, minScore);
         if (k <= 0 || k > K_MAX) {
             throw new IllegalArgumentException("[" + NAME + "] requires 0 < k <= " + K_MAX);
         }
@@ -108,33 +108,33 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
     }
 
     /**
-     * Builder method for max_distance
+     * Builder method for maxDistance
      *
-     * @param max_distance the max_distance threshold for the nearest neighbours
+     * @param maxDistance the maxDistance threshold for the nearest neighbours
      */
-    public KNNQueryBuilder maxDistance(Float max_distance) {
-        if (max_distance == null) {
-            throw new IllegalArgumentException("[" + NAME + "] requires max_distance to be set");
+    public KNNQueryBuilder maxDistance(Float maxDistance) {
+        if (maxDistance == null) {
+            throw new IllegalArgumentException("[" + NAME + "] requires maxDistance to be set");
         }
-        validateSingleQueryType(k, max_distance, min_score);
-        this.max_distance = max_distance;
+        validateSingleQueryType(k, maxDistance, minScore);
+        this.maxDistance = maxDistance;
         return this;
     }
 
     /**
-     * Builder method for min_score
+     * Builder method for minScore
      *
-     * @param min_score the min_score threshold for the nearest neighbours
+     * @param minScore the minScore threshold for the nearest neighbours
      */
-    public KNNQueryBuilder minScore(Float min_score) {
-        if (min_score == null) {
-            throw new IllegalArgumentException("[" + NAME + "] requires min_score to be set");
+    public KNNQueryBuilder minScore(Float minScore) {
+        if (minScore == null) {
+            throw new IllegalArgumentException("[" + NAME + "] requires minScore to be set");
         }
-        validateSingleQueryType(k, max_distance, min_score);
-        if (min_score <= 0) {
-            throw new IllegalArgumentException("[" + NAME + "] requires min_score greater than 0");
+        validateSingleQueryType(k, maxDistance, minScore);
+        if (minScore <= 0) {
+            throw new IllegalArgumentException("[" + NAME + "] requires minScore greater than 0");
         }
-        this.min_score = min_score;
+        this.minScore = minScore;
         return this;
     }
 
@@ -181,8 +181,8 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         this.k = k;
         this.filter = filter;
         this.ignoreUnmapped = false;
-        this.max_distance = null;
-        this.min_score = null;
+        this.maxDistance = null;
+        this.minScore = null;
     }
 
     public static void initialize(ModelDao modelDao) {
@@ -218,10 +218,10 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
                 ignoreUnmapped = in.readOptionalBoolean();
             }
             if (isClusterOnOrAfterMinRequiredVersion(KNNConstants.RADIAL_SEARCH_KEY)) {
-                max_distance = in.readOptionalFloat();
+                maxDistance = in.readOptionalFloat();
             }
             if (isClusterOnOrAfterMinRequiredVersion(KNNConstants.RADIAL_SEARCH_KEY)) {
-                min_score = in.readOptionalFloat();
+                minScore = in.readOptionalFloat();
             }
         } catch (IOException ex) {
             throw new RuntimeException("[KNN] Unable to create KNNQueryBuilder", ex);
@@ -233,8 +233,8 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         List<Object> vector = null;
         float boost = AbstractQueryBuilder.DEFAULT_BOOST;
         Integer k = null;
-        Float max_distance = null;
-        Float min_score = null;
+        Float maxDistance = null;
+        Float minScore = null;
         QueryBuilder filter = null;
         String queryName = null;
         String currentFieldName = null;
@@ -264,9 +264,9 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
                         } else if (AbstractQueryBuilder.NAME_FIELD.match(currentFieldName, parser.getDeprecationHandler())) {
                             queryName = parser.text();
                         } else if (MAX_DISTANCE_FIELD.match(currentFieldName, parser.getDeprecationHandler())) {
-                            max_distance = (Float) NumberFieldMapper.NumberType.FLOAT.parse(parser.objectBytes(), false);
+                            maxDistance = (Float) NumberFieldMapper.NumberType.FLOAT.parse(parser.objectBytes(), false);
                         } else if (MIN_SCORE_FIELD.match(currentFieldName, parser.getDeprecationHandler())) {
-                            min_score = (Float) NumberFieldMapper.NumberType.FLOAT.parse(parser.objectBytes(), false);
+                            minScore = (Float) NumberFieldMapper.NumberType.FLOAT.parse(parser.objectBytes(), false);
                         } else {
                             throw new ParsingException(
                                 parser.getTokenLocation(),
@@ -296,7 +296,7 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
             }
         }
 
-        validateSingleQueryType(k, max_distance, min_score);
+        validateSingleQueryType(k, maxDistance, minScore);
 
         KNNQueryBuilder knnQueryBuilder = new KNNQueryBuilder(fieldName, ObjectsToFloats(vector)).filter(filter)
             .ignoreUnmapped(ignoreUnmapped)
@@ -305,10 +305,10 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
 
         if (k != null) {
             knnQueryBuilder.k(k);
-        } else if (max_distance != null) {
-            knnQueryBuilder.maxDistance(max_distance);
-        } else if (min_score != null) {
-            knnQueryBuilder.minScore(min_score);
+        } else if (maxDistance != null) {
+            knnQueryBuilder.maxDistance(maxDistance);
+        } else if (minScore != null) {
+            knnQueryBuilder.minScore(minScore);
         }
 
         return knnQueryBuilder;
@@ -324,10 +324,10 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
             out.writeOptionalBoolean(ignoreUnmapped);
         }
         if (isClusterOnOrAfterMinRequiredVersion(KNNConstants.RADIAL_SEARCH_KEY)) {
-            out.writeOptionalFloat(max_distance);
+            out.writeOptionalFloat(maxDistance);
         }
         if (isClusterOnOrAfterMinRequiredVersion(KNNConstants.RADIAL_SEARCH_KEY)) {
-            out.writeOptionalFloat(min_score);
+            out.writeOptionalFloat(minScore);
         }
     }
 
@@ -350,11 +350,11 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
     }
 
     public float getMaxDistance() {
-        return this.max_distance;
+        return this.maxDistance;
     }
 
     public float getMinScore() {
-        return this.min_score;
+        return this.minScore;
     }
 
     public QueryBuilder getFilter() {
@@ -385,14 +385,14 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         if (filter != null) {
             builder.field(FILTER_FIELD.getPreferredName(), filter);
         }
-        if (max_distance != null) {
-            builder.field(MAX_DISTANCE_FIELD.getPreferredName(), max_distance);
+        if (maxDistance != null) {
+            builder.field(MAX_DISTANCE_FIELD.getPreferredName(), maxDistance);
         }
         if (ignoreUnmapped) {
             builder.field(IGNORE_UNMAPPED_FIELD.getPreferredName(), ignoreUnmapped);
         }
-        if (min_score != null) {
-            builder.field(MIN_SCORE_FIELD.getPreferredName(), min_score);
+        if (minScore != null) {
+            builder.field(MIN_SCORE_FIELD.getPreferredName(), minScore);
         }
         printBoostAndQueryName(builder);
         builder.endObject();
@@ -436,18 +436,18 @@ public class KNNQueryBuilder extends AbstractQueryBuilder<KNNQueryBuilder> {
         // Currently, k-NN supports distance and score types radial search
         // We need transform distance/score to right type of engine required radius.
         Float radius = null;
-        if (this.max_distance != null) {
-            if (this.max_distance < 0 && SpaceType.INNER_PRODUCT.equals(spaceType) == false) {
+        if (this.maxDistance != null) {
+            if (this.maxDistance < 0 && SpaceType.INNER_PRODUCT.equals(spaceType) == false) {
                 throw new IllegalArgumentException("[" + NAME + "] requires distance to be non-negative for space type: " + spaceType);
             }
-            radius = knnEngine.distanceToRadialThreshold(this.max_distance, spaceType);
+            radius = knnEngine.distanceToRadialThreshold(this.maxDistance, spaceType);
         }
 
-        if (this.min_score != null) {
-            if (this.min_score > 1 && SpaceType.INNER_PRODUCT.equals(spaceType) == false) {
+        if (this.minScore != null) {
+            if (this.minScore > 1 && SpaceType.INNER_PRODUCT.equals(spaceType) == false) {
                 throw new IllegalArgumentException("[" + NAME + "] requires score to be in the range (0, 1] for space type: " + spaceType);
             }
-            radius = knnEngine.scoreToRadialThreshold(this.min_score, spaceType);
+            radius = knnEngine.scoreToRadialThreshold(this.minScore, spaceType);
         }
 
         if (fieldDimension != vector.length) {